+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog for Critical Maps iOS
 ### Added
 - Message Notification Bubble
 - Swiftformat to the build phases
+- Network activity indicator support
 - French localisation. Thanks Alban!
 
 ### Updated